refactor: FrontendCodeBuild를 cicd 폴더 아래로 이동
